+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 Format is based on [Keep a Changelog] (https://keepachangelog.com/en/1.0.0/).
 Versionning adheres to [Semantic Versioning] (https://semver.org/spec/v2.0.0.html)
 
+## [Unrelease]
+### Modified
+- Break previous principles for verify functions. Verify functions are now
+  outputting errors in case of check fail and leave the end user the choice on
+  how to handle the situation in its unit testing code (for example t.Errorf or
+  t.Fatalf...)
+- Switch to github.com/pirmd/text v0.4.0 that offers better diff algorithm and
+  color diff formatting.
+- Use github.com/sanity-io/litter to provide readily usable pretty formatting
+  of compared values. 
+
+
 ## [0.3.0] - 2019-11-11
 ### Added
 - Add extended error message definition in MatchGolden